adding an idea for using SEXP as type
It would be necessary to implement every function for SEXP.

Idea: R interface works with the functions using SEXP and also the other functions using sexp arguments.

 Whereas the XPtr interface is using the sexp interface. However, the compiler decide which function has to be chosen. Thus, it is  possible to pass a SEXP variable to an external pointer function.

The advantage of the SEXP interface would be that it is not necessary to create sexp object for each SEXP variable. The intermediate results would be of type sexp. This results in a mix of SEXP and sexp variables.
Maybe the return value could be an SEXP object anyway. THus the conversion isn't necessary anymore......